requirements.txt
flask
google-generativeai
moviepy
opencv-python
requests
yt-dlp
schedule
instagrapi